requirements.txt
defusedxml==0.4.1
Django==1.7.4
django-admin-bootstrapped==2.3.3
django-allauth==0.19.1
django-bootstrap3==5.1.1
django-ckeditor==4.4.7
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-taggit==0.12.2
django-uuslug==1.0.3
oauthlib==0.7.2
psycopg2==2.6
python-slugify==0.1.0
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
sqlparse==0.1.14
Unidecode==0.4.17
